qdl: Allow multiple Sahara images
Some newer targets, such as Kaanapali, is requesting multiple images
from Sahara on the path to reach Firehose mode.

Extend the handling of the "programmer" argument to allow specifying a
comma-separated, colon-separated list of "<id>:<filename>" entries, to
specify the filenames for each of the image ids to be provided.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>