qdl: Support for "programmer archives"
Providing a long list of Sahara image ids and their respective payload
filename is cumbersome and error prone.

Introduce support for loading a "programmer archive", a CPIO archive
containing the list of Sahara images. The Sahara image id is read from
the filename of each file in the archive, and should be provided in the
form "<id>[:<filename>]".

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>