qdl: Allow skipping the firehose step by omitting xml files
Sometimes it is useful to run only the sahara step without running
firehose. Allow that by detecting if no xml files are passed.

This is mainly useful to mix and match qdl with stock tools like
QSaharaServer and fh_loader, in case some combination of these
behaves better for a given device.

Signed-off-by: Jerry Zhang <Jerry@skydio.com>